Added tests, docs and flake8
